Feature: perform the verify and write as background processes
Being able to cancel a long running process of remote files is important.
By backgrounding this, the GUI remains responsive and the user may cancel at any time.

The cancel process is done via polling instead of interrupt.
Both verify and write processes will completely finish operating on the current item before cancelling.
This is extremely useful in the write process because it effectively allows for cancelling a write and then later resuming.
The process does not resume, but because already downloaded files are not downloaded again, it functions effectively as a resume.

This is not the best implementation.
The ImportDataProcessor interface is not a good fit for being used as both a read and a write class if the write class is to be backgrounded.

The SwingWorker class only executes once, so everytime the write or verify processes are executed, a new class instance must be generated.
This includes when a process is cancelled.

The Verifier class was easier to implement, but I had to implement an abstract class to consistently extend the individual verifiers.