Feature: allow for continuing even when verification fails for remote bitstreams
When a bitstream fails, it is removed from the list (no meta data will be uploaded).
The directory is deleted on write (to prevent potential errors from having an empty folder).

When this feature is enabled, it is highly recommended to generate the flag CSV.
The flag CSV is placed in the output directory, so be sure to make the zip before generating or move the CSV before making the zip.