Update: implement basic mime-type validation for input CSV file
I confirmed that the program can load an excel spreadsheet and think everything is fine.
When the validate button is pressed, the program ends up locking/freezing.

Implement a very minimalistic sanity check to see if the input CSV file has a mime-type of 'text/plain'.
Other, related, checks are also performed.