Feature: add cleanup step for removing ignored folders or failed folders
By allowing for all failed items to be removed, the subset of the SAF batch of successful writes can still be used.
This is very helpful when the batch size is in the thousands.

This adds a failed status to the batch that designates if any given row is marked as having failed.
The ImportDataWriter has been broken up into an interface called ImportDataOperator allowing for the new cleanup class (ImportDataCleaner) to implement.

The new Cleanup button is only available when write has any failures or ignored failures.