Bugfix: cancelling a batch write doesn't cancel items being processed
This is a problem based on the design whereas the SwingWorker cancel operation exists at a higher level than where the items are processed in a loop.
An entire set of items will continue to be processed until completed even when cancelled.

The SwingWorkers may have different structures due to the templating design, so I used Reflection to pass along the isCancelled() method to the item.
This requires some significant changes but allows for the Item to check the isCancelled() status without knowing anything about the SwingWorker.

There are still some areas where this does not reach, but I decided this was good enough for now.
A refactor of the project is preferred but out of the scope of this fix.

ImportDataProcessorImpl provides a stub isCancelled() because it calls the writeItemSAF() method but is not currently a SwingWorker.