Update: fix bug with verifier sorting, finish implementing and use VerifierTableModel, and rename "Is Enabled" to "Activated"
Changing the sorting causes the verifier properties to not match to the correct row, breaking the enable/disabled toggle.

I noticed that I overlooked the existence of VerifierTableModel during the refactor.
This class should be completed and the verififier properties storage is redesigned to utilize this.

Rename "Is Enabled" to "Activated" to match the naming scheme that was already provided in the VerifierTableModel.