Bugfix: checkboxes associated with remote files are not always toggling
The problem is that I used a mouse listener when it should instead be an action listener.
The mouse listener would not fire when space is clicked so the field ends up getting a checkbox selection or de-selection without performing the additional required operations.
An action listener allows for space bar to toggle the checkbox.