Update: detect 'text/html' for bitstreams and treat as a failure
The bitstreams should never be 'text/html' and are most often 'application/pdf'.
Often times when a 'text/html' file is returned, the website is returning a 404 type of response but actually returns a 200 (quite incorrectly).
The files could then get stored as document-1.pdf but are actually document-1.html.

Mark cases where 'text/html' is returned from the remote file bitstream as write failures.